# server/openapi.yaml
openapi: 3.0.3
info:
  title: School Canteen API
  version: 1.0.0

paths:
  /menu:
    get:
      summary: List all menu items
      responses:
        '200':
          description: Array of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItem'

  /menu/{itemId}:
    get:
      summary: Get a single menu item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A menu item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '404':
          $ref: '#/components/responses/NotFound'

  /orders:
    post:
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /orders/history:
    get:
      summary: Get order history
      responses:
        '200':
          description: Array of past orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

components:
  schemas:
    MenuItem:
      type: object
      required: [id, name, description, price, category, emoji, customizations]
      properties:
        id:
          type: string
          example: m1
        name:
          type: string
          example: Grilled Chicken Rice
        description:
          type: string
        price:
          type: number
          format: float
          example: 8.50
        category:
          type: string
          enum: [Mains, Snacks, Beverages, Desserts]
        emoji:
          type: string
          example: 🍗
        customizations:
          type: array
          items:
            type: string
          example: ['Extra spicy', 'No onions']

    OrderItem:
      type: object
      required: [id, menuItemName, quantity, price, customizations, emoji]
      properties:
        id:
          type: string
        menuItemName:
          type: string
        quantity:
          type: integer
          minimum: 1
        price:
          type: number
          format: float
        customizations:
          type: array
          items:
            type: string
        emoji:
          type: string

    Order:
      type: object
      required: [id, date, items, total, status, estimatedTime]
      properties:
        id:
          type: string
          example: ORD-8492
        date:
          type: string
          example: Today, 10:30 AM
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total:
          type: number
          format: float
          example: 12.00
        status:
          type: string
          enum: [Confirmed, Preparing, Ready, Picked Up]
        estimatedTime:
          type: string
          example: 12:30 PM

    PlaceOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [itemId, quantity]
            properties:
              itemId:
                type: string
                example: m1
              quantity:
                type: integer
                minimum: 1
                example: 1
              customizations:
                type: array
                items:
                  type: string
                example: ['Extra spicy']
        pickupTime:
          type: string
          example: 12:30 PM

    Error:
      type: object
      properties:
        error:
          type: string

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'